+++ src/bindings/bnd_material.h
@@ -92,13 +92,13 @@ public:
   //public DocObjects.Texture[] GetTextures()
   bool Supported() const;
   //public void SynchronizeLegacyMaterial()
-  //public Rhino.Display.Color4f BaseColor{ get; set; }
   BND_Color4f BaseColor() const { return ON_4fColor_to_Binding(m_material->PhysicallyBased()->BaseColor()); }
   void SetBaseColor(const BND_Color4f& c) const { return m_material->PhysicallyBased()->SetBaseColor(Binding_to_ON_4fColor(c)); }
   BND_Color4f EmissionColor() const { return ON_4fColor_to_Binding(m_material->PhysicallyBased()->Emission()); }
   void SetEmissionColor(const BND_Color4f& c) const { return m_material->PhysicallyBased()->SetEmission(Binding_to_ON_4fColor(c)); }
+  BND_Color4f SubsurfaceScatteringColor() const { return ON_4fColor_to_Binding(m_material->PhysicallyBased()->SubsurfaceScatteringColor()); }
+  void SetSubsurfaceScatteringColor(const BND_Color4f& c) const { return m_material->PhysicallyBased()->SetSubsurfaceScatteringColor(Binding_to_ON_4fColor(c)); }
   //public BRDFs BRDF{ get; set; }
-  //public Rhino.Display.Color4f SubsurfaceScatteringColor{ get; set; }
   double Subsurface() const { return m_material->PhysicallyBased()->Subsurface(); }
   void SetSubsurface(double s) { m_material->PhysicallyBased()->SetSubsurface(s); }
   double SubsurfaceScatteringRadius() const { return m_material->PhysicallyBased()->SubsurfaceScatteringRadius(); }
@@ -131,6 +131,5 @@ public:
   void SetOpacity(double o) { m_material->PhysicallyBased()->SetOpacity(o); }
   double OpacityRoughness() const { return m_material->PhysicallyBased()->OpacityRoughness(); }
   void SetOpacityRoughness(double o) { m_material->PhysicallyBased()->SetOpacityRoughness(o); }
-  //public Rhino.Display.Color4f Emission{ get; set; }
 };